Randomize button works now
The randomize button now works.  I implemented the click event for the randomize button.  Each time it is pressed a new random image is generated and stored in the images folder.  I added a counter so it keeps all of the images and simply adds a counter to the end of the name starting at 0.  They will be overwritten when the program is run again so deleting is not necessary although I might add it to the on close event just to be tidy.  The image path is now relative and should work on different computers.